+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `dx catalog entities create/update/upsert`: Add support for assigning owner users by email with `--owner-user-emails`.
+
 ## 0.3.6 - 2026-05-14
 
 ### Added